Address Pietro's feedback
