+++ themes/ferrocene/static/fonts/raleway-700-italic.woff2
